Renaming hydra to cli
